requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
asyncpg==0.30.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
iso8601==2.1.0
multidict==6.5.0
propcache==0.3.2
pypika-tortoise==0.6.1
pytz==2025.2
tortoise-orm==0.25.1
typing_extensions==4.14.0
yarl==1.20.1
